inspektor.commands: Move GithubChecker to the github command
This way we nicely encapsulate github related functionality.
